Conflicts:
	cmake/Includes/SharedLibraries.cmake

+++ modules/behaviour/planning/WalkPathPlanner/src/WalkPathPlanner.h
@@ -79,7 +79,7 @@ namespace modules {
                     //vector<float> ballApproachAngle;
                     //vector<int> ballKickFoot;
                     float ballLineupDistance;
-                    int ballLineupMinDistance;
+                    float ballLineupMinDistance;
 
                     //extra config options
                     bool useAvoidance;
